populate_metabase_emplois: Move populate_table to metabase.db
It makes more sense to move this inner method, that should be unit
tested, with the utilities it heavily uses, since it does not depend on
anything specific inside the command.
